backend/requirements.txt
fastapi>=0.110,<0.116
uvicorn>=0.29,<0.31
psycopg[binary]>=3.1,<3.3
psycopg_pool>=3.2,<3.3
python-dotenv>=1.0,<2.0
passlib[bcrypt]>=1.7,<2.0
python-jose>=3.3,<4.0
aiosmtplib>=2.0,<4.0
email-validator>=2.1,<3.0
pydantic>=2.6,<2.9
orjson>=3.9,<3.11
boto3>=1.34,<1.35
pywebpush>=1.9,<3.0
cryptography>=41,<43
sentry-sdk[fastapi]>=2.5,<2.8
prometheus-client>=0.17,<0.21
